Changed Come bet to use a Point object instead of None or number.
