+++ crapssim/strategy/core.py
@@ -127,6 +127,7 @@ class BetIfTrue(Strategy):
         player
             The Player to add the bet for.
         """
+        print(player.bets_on_table, self.bet)
         if self.key(player):
             player.add_bet(self.bet)
